Start_application and stop_application are now model methods
