+++ remoteappmanager/static/js/home/views/application_list_view.js
@@ -1,40 +1,13 @@
 define([
-    'urlutils',
     'utils',
-    "dialogs",
     '../../components/vue/dist/vue.min',
-    "jsapi/v1/resources"
-], function (urlutils, utils, dialogs, Vue, resources) {
+], function (utils, Vue) {
     'use strict';
-
-    var Status = utils.Status;
 
     /* Create application_list ViewModel
     (will next be wrapped in a main ViewModel which will contain the
     applicationListView and the applicationView) */
     var ApplicationListView = Vue.extend({
-        methods: {
-            stop_application: function(index) {
-                var app_stopping = this.model.app_list[index];
-                app_stopping.status = Status.STOPPING;
-
-                var url_id = app_stopping.app_data.container.url_id;
-
-                resources.Container.delete(url_id)
-                .done(function() {
-                    this.model.update_idx(index)
-                    .fail(function(error) {
-                        app_stopping.status = Status.STOPPED;
-                        dialogs.webapi_error_dialog(error);
-                    });
-                }.bind(this))
-                .fail(function(error) {
-                    app_stopping.status = Status.STOPPED;
-                    dialogs.webapi_error_dialog(error);
-                });
-            }
-        },
-
         filters: utils.filters
     });
