+++ remoteappmanager/templates/home.html
@@ -53,7 +53,7 @@
 
         <button class="stop-button"
                 v-if="app.status === 'RUNNING'"
-                @click="stop_application(index)"
+                @click="model.stop_application(index)"
                 :disabled="app | is_stopping">
           <i class="fa fa-times"></i>
         </button>
@@ -111,7 +111,7 @@
           <!-- Start Button -->
           <div class="box-footer">
             <button class="btn btn-primary pull-right start-button"
-                    @click="start_application()"
+                    @click="model.start_application()"
                     :disabled="current_app | is_starting">
               Start
             </button>
